Allow user to pass in database credentials file
Database credentials such as the host, port, user, and database
name can be stored in a JSON file and passed to the program as a
command line argument. The program will use whats in this JSON file
to set up a connection to the appropriate database. A usage error
message is displayed if no JSON file is provided.